# core/backend/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Lab-Core Backend API
  version: 0.2.0
  description: |
    Lab-Core backend API の正規仕様です。
    実行単位は Job ではなく Operation を正とし、旧 `/api/jobs` と旧操作 API は含みません。
servers:
  - url: /
...
paths:
...
  /api/operations/{operationId}:
    get:
      tags: [Operations]
      operationId: getOperation
      summary: Operation 詳細を取得
      parameters:
        - $ref: '#/components/parameters/OperationId'
      responses:
        "200":
          description: Operation 詳細
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OperationDetail'
        "404":
          $ref: '#/components/responses/ErrorResponse'
  /api/operations/{operationId}/cancel:
    post:
      tags: [Operations]
      operationId: cancelOperation
      summary: queued Operation を cancel
      parameters:
        - $ref: '#/components/parameters/OperationId'
      responses:
        "202":
          description: cancel 後の Operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OperationDetail'
        "404":
          $ref: '#/components/responses/ErrorResponse'
        "409":
          $ref: '#/components/responses/ErrorResponse'
  /api/operations/{operationId}/retry:
    post:
      tags: [Operations]
      operationId: retryOperation
      summary: failed/interrupted Operation を新しい Operation として retry
      parameters:
        - $ref: '#/components/parameters/OperationId'
      responses:
        "202":
          description: 新しい retry Operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OperationDetail'
        "404":
          $ref: '#/components/responses/ErrorResponse'
        "409":
          $ref: '#/components/responses/ErrorResponse'
  /api/operations/{operationId}/logs:
    get:
      tags: [Operations]
      operationId: getOperationLogs
      summary: Operation Log を取得
      parameters:
        - $ref: '#/components/parameters/OperationId'
        - in: query
          name: after
          schema:
            type: integer
            minimum: 0
          required: false
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
            maximum: 1000
            default: 1000
          required: false
        - in: query
          name: tail
          schema:
            type: integer
            minimum: 1
            maximum: 1000
          required: false
        - in: query
          name: stepId
          schema:
            type: string
          required: false
      responses:
        "200":
          description: Operation Log
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OperationLogListResponse'
        "404":
          $ref: '#/components/responses/ErrorResponse'
  /api/operations/{operationId}/logs/stream:
    get:
      tags: [Operations]
      operationId: streamOperationLogs
      summary: Operation Log / Step / Operation event を SSE で購読
      parameters:
        - $ref: '#/components/parameters/OperationId'
      responses:
        "200":
          description: SSE stream
          content:
            text/event-stream:
              schema:
                type: string
...
components:
  parameters:
...
    OperationId:
      name: operationId
      in: path
      required: true
      schema:
        type: string
  responses:
    ErrorResponse:
      description: 標準エラーレスポンス
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: object
              additionalProperties: true
              nullable: true
            requestId:
              type: string
          required: [code, message, requestId]
      required: [error]
...
    Operation:
      type: object
      properties:
        operationId:
          type: string
        applicationId:
          type: string
        type:
          type: string
          enum: [deploy, restart, stop, resume, rebuild, update-check, update, rollback, delete]
        status:
          type: string
          enum: [queued, running, succeeded, failed, cancelled, interrupted]
        currentStepId:
          type: string
          nullable: true
        currentStepName:
          type: string
          nullable: true
        currentStepOrder:
          type: integer
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        startedAt:
          type: string
          format: date-time
          nullable: true
        finishedAt:
          type: string
          format: date-time
          nullable: true
        parameters:
          type: object
          additionalProperties: true
        result:
          type: object
          additionalProperties: true
          nullable: true
        error:
          type: object
          nullable: true
        retryOfOperationId:
          type: string
          nullable: true
        logsAvailable:
          type: boolean
        canCancel:
          type: boolean
        canRetry:
          type: boolean
        links:
          type: object
          properties:
            self:
              type: string
            logs:
              type: string
            logStream:
              type: string
          required: [self, logs, logStream]
      required:
        - operationId
        - applicationId
        - type
        - status
        - createdAt
        - updatedAt
        - parameters
        - logsAvailable
        - canCancel
        - canRetry
        - links
    OperationStep:
      type: object
      properties:
        stepId:
          type: string
        operationId:
          type: string
        stepOrder:
          type: integer
        name:
          type: string
        status:
          type: string
          enum: [pending, running, succeeded, failed, skipped]
        startedAt:
          type: string
          format: date-time
          nullable: true
        updatedAt:
          type: string
          format: date-time
        finishedAt:
          type: string
          format: date-time
          nullable: true
        message:
          type: string
          nullable: true
        errorCode:
          type: string
          nullable: true
        details:
          type: object
          additionalProperties: true
          nullable: true
      required: [stepId, operationId, stepOrder, name, status, updatedAt]
    OperationDetail:
      allOf:
        - $ref: '#/components/schemas/Operation'
        - type: object
          properties:
            steps:
              type: array
              items:
                $ref: '#/components/schemas/OperationStep'
          required: [steps]
    OperationLogItem:
      type: object
      properties:
        operationId:
          type: string
        sequence:
          type: integer
        stepId:
          type: string
          nullable: true
        stream:
          type: string
          enum: [stdout, stderr, system]
        line:
          type: string
        createdAt:
          type: string
          format: date-time
      required: [operationId, sequence, stepId, stream, line, createdAt]
    OperationLogListResponse:
      type: object
      properties:
        operationId:
          type: string
        logsAvailable:
          type: boolean
        items:
          type: array
          items:
            $ref: '#/components/schemas/OperationLogItem'
        nextAfter:
          type: integer
          nullable: true
        hasMore:
          type: boolean
      required: [operationId, logsAvailable, items, nextAfter, hasMore]
...